fix configuration management for package gdb (#2163)
