+++ contrib/drivers/oracle/oracle_init_test.go
@@ -9,13 +9,14 @@ package oracle_test
 import (
 	"context"
 	"fmt"
+	"strings"
+
 	"github.com/gogf/gf/v2/container/garray"
 	"github.com/gogf/gf/v2/database/gdb"
 	"github.com/gogf/gf/v2/frame/g"
 	"github.com/gogf/gf/v2/os/gtime"
 	"github.com/gogf/gf/v2/test/gtest"
 	_ "github.com/sijms/go-ora/v2"
-	"strings"
 )
 
 var (
@@ -61,8 +62,9 @@ func init() {
 	nodeLink := gdb.ConfigNode{
 		Type: TestDbType,
 		Name: TestDbName,
-		Link: fmt.Sprintf("%s://%s:%s@%s:%s/%s",
-			TestDbType, TestDbUser, TestDbPass, TestDbIP, TestDbPort, TestDbName),
+		Link: fmt.Sprintf("%s:%s:%s@tcp(%s:%s)/%s",
+			TestDbType, TestDbUser, TestDbPass, TestDbIP, TestDbPort, TestDbName,
+		),
 	}
 
 	nodeErr := gdb.ConfigNode{
