+++ database/gdb/gdb_driver_wrapper_db.go
@@ -13,9 +13,7 @@ import (
 
 	"github.com/gogf/gf/v2/errors/gcode"
 	"github.com/gogf/gf/v2/errors/gerror"
-	"github.com/gogf/gf/v2/text/gregex"
 	"github.com/gogf/gf/v2/text/gstr"
-	"github.com/gogf/gf/v2/util/gconv"
 	"github.com/gogf/gf/v2/util/gutil"
 )
 
@@ -27,9 +25,6 @@ type DriverWrapperDB struct {
 // Open creates and returns an underlying sql.DB object for pgsql.
 // https://pkg.go.dev/github.com/lib/pq
 func (d *DriverWrapperDB) Open(config *ConfigNode) (db *sql.DB, err error) {
-	if config.Link != "" {
-		config = parseConfigNodeLink(config)
-	}
 	return d.DB.Open(config)
 }
 
@@ -82,42 +77,4 @@ func (d *DriverWrapperDB) TableFields(ctx context.Context, table string, schema 
 		fields = value.(map[string]*TableField)
 	}
 	return
-}
-
-func parseConfigNodeLink(node *ConfigNode) *ConfigNode {
-	var match []string
-	if node.Link != "" {
-		match, _ = gregex.MatchString(linkPattern, node.Link)
-		if len(match) > 5 {
-			node.Type = match[1]
-			node.User = match[2]
-			node.Pass = match[3]
-			node.Protocol = match[4]
-			array := gstr.Split(match[5], ":")
-			if len(array) == 2 {
-				node.Host = array[0]
-				node.Port = array[1]
-				node.Name = match[6]
-			} else {
-				node.Name = match[5]
-			}
-			if len(match) > 6 {
-				node.Extra = match[7]
-			}
-			node.Link = ""
-		}
-	}
-	if node.Extra != "" {
-		if m, _ := gstr.Parse(node.Extra); len(m) > 0 {
-			_ = gconv.Struct(m, &node)
-		}
-	}
-	// Default value checks.
-	if node.Charset == "" {
-		node.Charset = defaultCharset
-	}
-	if node.Protocol == "" {
-		node.Protocol = defaultProtocol
-	}
-	return node
 }
