+++ version.go
@@ -1,4 +1,4 @@
 package gf
 
-const VERSION = "v2.2.0-beta"
+const VERSION = "v2.2.0-beta2"
 const AUTHORS = "john<john@goframe.org>"
